Move internal/pkg/config/types.go to public /config package
Fixes: https://github.com/edgexfoundry/go-mod-bootstrap/issues/1
